refactor: move custom error to subdir
